# backend/tests/fixtures/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Evidence": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "quote": {
            "default": "",
            "title": "Quote",
            "type": "string"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "title": "Evidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemCreate": {
        "properties": {
          "description": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "dueDate": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "owner": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "rationale": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "type": {
            "enum": [
              "IDEA",
              "DECISION",
              "ACTION",
              "QUESTION"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "description"
        ],
        "title": "ItemCreate",
        "type": "object"
      },
      "ItemDetail": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/KnowledgeItem"
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeItem"
            },
            "title": "Related",
            "type": "array"
          },
          "similar": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeItem"
            },
            "title": "Similar",
            "type": "array"
          }
        },
        "required": [
          "item",
          "related"
        ],
        "title": "ItemDetail",
        "type": "object"
      },
...
      "ItemTopicUpdate": {
        "properties": {
          "topicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicid"
          }
        },
        "title": "ItemTopicUpdate",
        "type": "object"
      },
      "ItemUpdate": {
        "description": "Partial edit: omitted fields are left alone; an explicit null/blank clears owner, due date or rationale.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dueDate": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "owner": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "rationale": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "type": {
            "anyOf": [
              {
                "enum": [
                  "IDEA",
                  "DECISION",
                  "ACTION",
                  "QUESTION"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "ItemUpdate",
        "type": "object"
      },
      "ItemsTopicBulkUpdate": {
        "properties": {
          "itemIds": {
            "items": {
              "type": "string"
            },
            "title": "Itemids",
            "type": "array"
          },
          "topicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicid"
          }
        },
        "required": [
          "itemIds"
        ],
        "title": "ItemsTopicBulkUpdate",
        "type": "object"
      },
      "KnowledgeItem": {
        "properties": {
          "confidence": {
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "dueDateSourceText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedatesourcetext"
          },
          "effectivePriority": {
            "anyOf": [
              {
                "enum": [
                  "CRITICAL",
                  "MAJOR",
                  "MINOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effectivepriority"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manualOverride": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualPriorityOverride"
              },
              {
                "type": "null"
              }
            ]
          },
          "meetingId": {
            "title": "Meetingid",
            "type": "string"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "title": "Notes",
            "type": "array"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "relatedIds": {
            "items": {
              "type": "string"
            },
            "title": "Relatedids",
            "type": "array"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "stakeholders": {
            "items": {
              "type": "string"
            },
            "title": "Stakeholders",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "topicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicid"
          },
          "topicName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicname"
          },
          "type": {
            "enum": [
              "IDEA",
              "DECISION",
              "ACTION",
              "QUESTION"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "description",
          "status",
          "confidence",
          "evidence",
          "meetingId"
        ],
        "title": "KnowledgeItem",
        "type": "object"
      },
      "ManualPriorityOverride": {
        "properties": {
          "overriddenAt": {
            "title": "Overriddenat",
            "type": "string"
          },
          "priority": {
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR"
            ],
            "title": "Priority",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "priority",
          "overriddenAt"
        ],
        "title": "ManualPriorityOverride",
        "type": "object"
      },
...
      "NoteCreate": {
        "properties": {
          "body": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "NoteCreate",
        "type": "object"
      },
      "PriorityOverrideUpdate": {
        "properties": {
          "priority": {
            "anyOf": [
              {
                "enum": [
                  "CRITICAL",
                  "MAJOR",
                  "MINOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "PriorityOverrideUpdate",
        "type": "object"
      },
...
      "StatusUpdate": {
        "properties": {
          "status": {
            "enum": [
              "Open",
              "Closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusUpdate",
        "type": "object"
      },
      "TagCreate": {
        "properties": {
          "tag": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "tag"
        ],
        "title": "TagCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "second-brain backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/items": {
      "get": {
        "operationId": "get_items_api_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "IDEA",
                    "DECISION",
                    "ACTION",
                    "QUESTION"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "CRITICAL",
                    "MAJOR",
                    "MINOR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Open",
                    "Closed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeItem"
                  },
                  "title": "Response Get Items Api Items Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Items"
      }
    },
    "/api/items/topic": {
      "patch": {
        "operationId": "set_items_topic_api_items_topic_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemsTopicBulkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeItem"
                  },
                  "title": "Response Set Items Topic Api Items Topic Patch",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Items Topic"
      }
    },
    "/api/items/{item_id}": {
      "delete": {
        "operationId": "delete_item_api_items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Item"
      },
      "get": {
        "operationId": "get_item_api_items__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Item"
      },
      "patch": {
        "operationId": "update_item_api_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Item"
      }
    },
    "/api/items/{item_id}/notes": {
      "post": {
        "operationId": "add_item_note_api_items__item_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Item Note"
      }
    },
    "/api/items/{item_id}/notes/{note_id}": {
      "delete": {
        "operationId": "delete_item_note_api_items__item_id__notes__note_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Item Note"
      },
      "patch": {
        "operationId": "update_item_note_api_items__item_id__notes__note_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Item Note"
      }
    },
    "/api/items/{item_id}/priority-override": {
      "patch": {
        "operationId": "set_item_priority_override_api_items__item_id__priority_override_patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityOverrideUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Item Priority Override"
      }
    },
    "/api/items/{item_id}/status": {
      "patch": {
        "operationId": "set_item_status_api_items__item_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Item Status"
      }
    },
    "/api/items/{item_id}/tags": {
      "delete": {
        "operationId": "remove_item_tag_api_items__item_id__tags_delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Tag",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Item Tag"
      },
      "post": {
        "operationId": "add_item_tag_api_items__item_id__tags_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Item Tag"
      }
    },
    "/api/items/{item_id}/topic": {
      "patch": {
        "operationId": "set_item_topic_api_items__item_id__topic_patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemTopicUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Item Topic"
      }
    },
...
    "/api/topics/{topic_id}/items": {
      "post": {
        "operationId": "create_topic_item_api_topics__topic_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Topic Item"
      }
    },
...
  }
}